minor adjustment

+++ bolt/bolt.h
@@ -70,7 +70,7 @@ uint32_t bolt_read(uint8_t* out_data);
  * @param   len:  the number of bytes to write (max. length: BOLT_MAX_MSG_LEN)
  * @return  true if the operation was successful, false otherwise
  */
-bool bolt_write(uint8_t* data, uint32_t len);
+bool bolt_write(uint8_t* data, uint16_t len);
 
 /*
  * get the write counter
